Refactor application startup process in `main.py` to consolidate production and development modes into a single `run_app` function. Enhanced logging for both environments and improved handling of application settings. Updated `gradio_app.py` to streamline imports and removed obsolete code related to chat streaming. Introduced `MCPFactory` class in `server.py` to modularize MCP server setup and configuration, improving code organization and readability.
