+++ src/api/__init__.py
@@ -1,0 +1,5 @@
+"""
+Module API pour l'agent IA d'inclusion sociale.
+
+Ce module contient les routes et endpoints de l'API REST.
+"""
