+++ src/ui/__init__.py
@@ -1,0 +1,5 @@
+"""
+Module UI pour l'agent IA d'inclusion sociale.
+
+Ce module contient les composants d'interface utilisateur Gradio.
+"""
